perf(results): extract results markers out of the FastAPI sidecar module
system_controller and records_manager imported aiperf.kubernetes.results_sidecar
for write_ready_marker / write_processing_marker / CHECKPOINTS_DIR_NAME. That
module is the sidecar HTTP server, so the import pulled uvicorn, FastAPI,
starlette, aiofiles and click (~100 modules, ~560 ms) into two service processes
of every local run purely for a filesystem contract.

The marker half of the module is pure stdlib plus a lazily imported orjson, so
move it to aiperf.common.results_markers and have results_sidecar re-export
every name for its ~40 existing importers. The three local-path call sites now
import from the neutral module.
